+++ doc/CHANGELOG.md
@@ -1,5 +1,13 @@
 ### Changelog
 
+## 2.1.0
+ * Diagram context save and load
+   * store a `vr2ctx` file (json format) with the paths to oreqm file(s), selection criteria and settings afecting rendering
+     Use the "Save diagram context..." and "Load diagram context...". The vr2ctx file uses both absolute paths and relative paths
+     the vr2ctx file. This allows copying of oreqm files as well as vr2ctx file to another user, who can then open exactly the same
+     view.
+     Caveat: Settings in program are updated by the vr2ctx file, but are not saved. To do this enter the settings dialog and exit with [OK]
+
 ## 2.0.3
  * Display more specobjects fields
    * securityrationale